pyproject.toml
[project]
name = "ai-note"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "black>=25.11.0",
    "fastapi>=0.121.2",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
